+++ packages/stream_core_flutter/lib/src/theme/primitives/stream_icons.g.theme.dart
@@ -38,9 +38,15 @@ mixin _$StreamIcons {
       audio: t < 0.5 ? a.audio : b.audio,
       bell: t < 0.5 ? a.bell : b.bell,
       bellOff: t < 0.5 ? a.bellOff : b.bellOff,
+      blurFill: t < 0.5 ? a.blurFill : b.blurFill,
       bolt: t < 0.5 ? a.bolt : b.bolt,
+      boltFill: t < 0.5 ? a.boltFill : b.boltFill,
       camera: t < 0.5 ? a.camera : b.camera,
+      cameraFlipFill: t < 0.5 ? a.cameraFlipFill : b.cameraFlipFill,
       cameraLarge: t < 0.5 ? a.cameraLarge : b.cameraLarge,
+      captionFill: t < 0.5 ? a.captionFill : b.captionFill,
+      caretDown: t < 0.5 ? a.caretDown : b.caretDown,
+      caretUp: t < 0.5 ? a.caretUp : b.caretUp,
       checkmark: t < 0.5 ? a.checkmark : b.checkmark,
       checks: t < 0.5 ? a.checks : b.checks,
       chevronDown: t < 0.5 ? a.chevronDown : b.chevronDown,
@@ -50,11 +56,14 @@ mixin _$StreamIcons {
       clock: t < 0.5 ? a.clock : b.clock,
       command: t < 0.5 ? a.command : b.command,
       copy: t < 0.5 ? a.copy : b.copy,
+      copyFill: t < 0.5 ? a.copyFill : b.copyFill,
+      darkMode: t < 0.5 ? a.darkMode : b.darkMode,
       delete: t < 0.5 ? a.delete : b.delete,
       download: t < 0.5 ? a.download : b.download,
       edit: t < 0.5 ? a.edit : b.edit,
       emoji: t < 0.5 ? a.emoji : b.emoji,
       emojiAdd: t < 0.5 ? a.emojiAdd : b.emojiAdd,
+      emojiAddFill: t < 0.5 ? a.emojiAddFill : b.emojiAddFill,
       exclamationCircle: t < 0.5 ? a.exclamationCircle : b.exclamationCircle,
       exclamationCircleFill: t < 0.5
           ? a.exclamationCircleFill
@@ -71,37 +80,59 @@ mixin _$StreamIcons {
       flag: t < 0.5 ? a.flag : b.flag,
       folder: t < 0.5 ? a.folder : b.folder,
       folderLarge: t < 0.5 ? a.folderLarge : b.folderLarge,
+      fullBlurFill: t < 0.5 ? a.fullBlurFill : b.fullBlurFill,
+      fullscreenFill: t < 0.5 ? a.fullscreenFill : b.fullscreenFill,
       gallery: t < 0.5 ? a.gallery : b.gallery,
+      gridFill: t < 0.5 ? a.gridFill : b.gridFill,
+      gridPixelFill: t < 0.5 ? a.gridPixelFill : b.gridPixelFill,
       image: t < 0.5 ? a.image : b.image,
       imageLarge: t < 0.5 ? a.imageLarge : b.imageLarge,
       info: t < 0.5 ? a.info : b.info,
+      language: t < 0.5 ? a.language : b.language,
       leave: t < 0.5 ? a.leave : b.leave,
+      leftToRight: t < 0.5 ? a.leftToRight : b.leftToRight,
+      lightMode: t < 0.5 ? a.lightMode : b.lightMode,
       link: t < 0.5 ? a.link : b.link,
       location: t < 0.5 ? a.location : b.location,
       lock: t < 0.5 ? a.lock : b.lock,
       megaphone: t < 0.5 ? a.megaphone : b.megaphone,
       mention: t < 0.5 ? a.mention : b.mention,
+      menu: t < 0.5 ? a.menu : b.menu,
       messageBubble: t < 0.5 ? a.messageBubble : b.messageBubble,
       messageBubbleFill: t < 0.5 ? a.messageBubbleFill : b.messageBubbleFill,
       messageBubbleLarge: t < 0.5 ? a.messageBubbleLarge : b.messageBubbleLarge,
       messageBubbles: t < 0.5 ? a.messageBubbles : b.messageBubbles,
+      messageBubblesFill: t < 0.5 ? a.messageBubblesFill : b.messageBubblesFill,
       messageBubblesLarge: t < 0.5
           ? a.messageBubblesLarge
           : b.messageBubblesLarge,
       minus: t < 0.5 ? a.minus : b.minus,
       minusCircle: t < 0.5 ? a.minusCircle : b.minusCircle,
       more: t < 0.5 ? a.more : b.more,
+      moreHorizontal: t < 0.5 ? a.moreHorizontal : b.moreHorizontal,
+      moreVerticalFill: t < 0.5 ? a.moreVerticalFill : b.moreVerticalFill,
       mute: t < 0.5 ? a.mute : b.mute,
       noSign: t < 0.5 ? a.noSign : b.noSign,
+      noSignFill: t < 0.5 ? a.noSignFill : b.noSignFill,
       notification: t < 0.5 ? a.notification : b.notification,
       pauseFill: t < 0.5 ? a.pauseFill : b.pauseFill,
+      phoneDownFill: t < 0.5 ? a.phoneDownFill : b.phoneDownFill,
       pin: t < 0.5 ? a.pin : b.pin,
+      pinFill: t < 0.5 ? a.pinFill : b.pinFill,
       pinLarge: t < 0.5 ? a.pinLarge : b.pinLarge,
+      pipFill: t < 0.5 ? a.pipFill : b.pipFill,
       playFill: t < 0.5 ? a.playFill : b.playFill,
       plus: t < 0.5 ? a.plus : b.plus,
       poll: t < 0.5 ? a.poll : b.poll,
       pollLarge: t < 0.5 ? a.pollLarge : b.pollLarge,
+      presentDesktopFill: t < 0.5 ? a.presentDesktopFill : b.presentDesktopFill,
+      presentMobileFill: t < 0.5 ? a.presentMobileFill : b.presentMobileFill,
+      questionCircleFill: t < 0.5 ? a.questionCircleFill : b.questionCircleFill,
       quote: t < 0.5 ? a.quote : b.quote,
+      raiseHandFill: t < 0.5 ? a.raiseHandFill : b.raiseHandFill,
+      recordLibraryFill: t < 0.5 ? a.recordLibraryFill : b.recordLibraryFill,
+      recordingFill: t < 0.5 ? a.recordingFill : b.recordingFill,
+      recordingStopFill: t < 0.5 ? a.recordingStopFill : b.recordingStopFill,
       refresh: t < 0.5 ? a.refresh : b.refresh,
       reorder: t < 0.5 ? a.reorder : b.reorder,
       reply: t < 0.5 ? a.reply : b.reply,
@@ -109,9 +140,14 @@ mixin _$StreamIcons {
       save: t < 0.5 ? a.save : b.save,
       search: t < 0.5 ? a.search : b.search,
       send: t < 0.5 ? a.send : b.send,
+      settings: t < 0.5 ? a.settings : b.settings,
+      settingsFill: t < 0.5 ? a.settingsFill : b.settingsFill,
       share: t < 0.5 ? a.share : b.share,
       shield: t < 0.5 ? a.shield : b.shield,
       sidebar: t < 0.5 ? a.sidebar : b.sidebar,
+      slidersFill: t < 0.5 ? a.slidersFill : b.slidersFill,
+      starFill: t < 0.5 ? a.starFill : b.starFill,
+      statsFill: t < 0.5 ? a.statsFill : b.statsFill,
       stopFill: t < 0.5 ? a.stopFill : b.stopFill,
       thread: t < 0.5 ? a.thread : b.thread,
       threadFill: t < 0.5 ? a.threadFill : b.threadFill,
@@ -119,6 +155,7 @@ mixin _$StreamIcons {
       trophy: t < 0.5 ? a.trophy : b.trophy,
       unlock: t < 0.5 ? a.unlock : b.unlock,
       unpin: t < 0.5 ? a.unpin : b.unpin,
+      unpinFill: t < 0.5 ? a.unpinFill : b.unpinFill,
       unsave: t < 0.5 ? a.unsave : b.unsave,
       unsupportedAttachment: t < 0.5
           ? a.unsupportedAttachment
@@ -126,13 +163,19 @@ mixin _$StreamIcons {
       upload: t < 0.5 ? a.upload : b.upload,
       user: t < 0.5 ? a.user : b.user,
       userAdd: t < 0.5 ? a.userAdd : b.userAdd,
+      userAddFill: t < 0.5 ? a.userAddFill : b.userAddFill,
       userCheck: t < 0.5 ? a.userCheck : b.userCheck,
       userRemove: t < 0.5 ? a.userRemove : b.userRemove,
+      userRemoveFill: t < 0.5 ? a.userRemoveFill : b.userRemoveFill,
       users: t < 0.5 ? a.users : b.users,
+      usersFill: t < 0.5 ? a.usersFill : b.usersFill,
+      verifiedFill: t < 0.5 ? a.verifiedFill : b.verifiedFill,
       video: t < 0.5 ? a.video : b.video,
       videoFill: t < 0.5 ? a.videoFill : b.videoFill,
+      videoOffFill: t < 0.5 ? a.videoOffFill : b.videoOffFill,
       voice: t < 0.5 ? a.voice : b.voice,
       voiceFill: t < 0.5 ? a.voiceFill : b.voiceFill,
+      voiceOffFill: t < 0.5 ? a.voiceOffFill : b.voiceOffFill,
       xCircle: t < 0.5 ? a.xCircle : b.xCircle,
       xmark: t < 0.5 ? a.xmark : b.xmark,
       xmarkSmall: t < 0.5 ? a.xmarkSmall : b.xmarkSmall,
@@ -154,9 +197,15 @@ mixin _$StreamIcons {
     IconData? audio,
     IconData? bell,
     IconData? bellOff,
+    IconData? blurFill,
     IconData? bolt,
+    IconData? boltFill,
     IconData? camera,
+    IconData? cameraFlipFill,
     IconData? cameraLarge,
+    IconData? captionFill,
+    IconData? caretDown,
+    IconData? caretUp,
     IconData? checkmark,
     IconData? checks,
     IconData? chevronDown,
@@ -166,11 +215,14 @@ mixin _$StreamIcons {
     IconData? clock,
     IconData? command,
     IconData? copy,
+    IconData? copyFill,
+    IconData? darkMode,
     IconData? delete,
     IconData? download,
     IconData? edit,
     IconData? emoji,
     IconData? emojiAdd,
+    IconData? emojiAddFill,
     IconData? exclamationCircle,
     IconData? exclamationCircleFill,
     IconData? exclamationMarkFill,
@@ -181,35 +233,57 @@ mixin _$StreamIcons {
     IconData? flag,
     IconData? folder,
     IconData? folderLarge,
+    IconData? fullBlurFill,
+    IconData? fullscreenFill,
     IconData? gallery,
+    IconData? gridFill,
+    IconData? gridPixelFill,
     IconData? image,
     IconData? imageLarge,
     IconData? info,
+    IconData? language,
     IconData? leave,
+    IconData? leftToRight,
+    IconData? lightMode,
     IconData? link,
     IconData? location,
     IconData? lock,
     IconData? megaphone,
     IconData? mention,
+    IconData? menu,
     IconData? messageBubble,
     IconData? messageBubbleFill,
     IconData? messageBubbleLarge,
     IconData? messageBubbles,
+    IconData? messageBubblesFill,
     IconData? messageBubblesLarge,
     IconData? minus,
     IconData? minusCircle,
     IconData? more,
+    IconData? moreHorizontal,
+    IconData? moreVerticalFill,
     IconData? mute,
     IconData? noSign,
+    IconData? noSignFill,
     IconData? notification,
     IconData? pauseFill,
+    IconData? phoneDownFill,
     IconData? pin,
+    IconData? pinFill,
     IconData? pinLarge,
+    IconData? pipFill,
     IconData? playFill,
     IconData? plus,
     IconData? poll,
     IconData? pollLarge,
+    IconData? presentDesktopFill,
+    IconData? presentMobileFill,
+    IconData? questionCircleFill,
     IconData? quote,
+    IconData? raiseHandFill,
+    IconData? recordLibraryFill,
+    IconData? recordingFill,
+    IconData? recordingStopFill,
     IconData? refresh,
     IconData? reorder,
     IconData? reply,
@@ -217,9 +291,14 @@ mixin _$StreamIcons {
     IconData? save,
     IconData? search,
     IconData? send,
+    IconData? settings,
+    IconData? settingsFill,
     IconData? share,
     IconData? shield,
     IconData? sidebar,
+    IconData? slidersFill,
+    IconData? starFill,
+    IconData? statsFill,
     IconData? stopFill,
     IconData? thread,
     IconData? threadFill,
@@ -227,18 +306,25 @@ mixin _$StreamIcons {
     IconData? trophy,
     IconData? unlock,
     IconData? unpin,
+    IconData? unpinFill,
     IconData? unsave,
     IconData? unsupportedAttachment,
     IconData? upload,
     IconData? user,
     IconData? userAdd,
+    IconData? userAddFill,
     IconData? userCheck,
     IconData? userRemove,
+    IconData? userRemoveFill,
     IconData? users,
+    IconData? usersFill,
+    IconData? verifiedFill,
     IconData? video,
     IconData? videoFill,
+    IconData? videoOffFill,
     IconData? voice,
     IconData? voiceFill,
+    IconData? voiceOffFill,
     IconData? xCircle,
     IconData? xmark,
     IconData? xmarkSmall,
@@ -260,9 +346,15 @@ mixin _$StreamIcons {
       audio: audio ?? _this.audio,
       bell: bell ?? _this.bell,
       bellOff: bellOff ?? _this.bellOff,
+      blurFill: blurFill ?? _this.blurFill,
       bolt: bolt ?? _this.bolt,
+      boltFill: boltFill ?? _this.boltFill,
       camera: camera ?? _this.camera,
+      cameraFlipFill: cameraFlipFill ?? _this.cameraFlipFill,
       cameraLarge: cameraLarge ?? _this.cameraLarge,
+      captionFill: captionFill ?? _this.captionFill,
+      caretDown: caretDown ?? _this.caretDown,
+      caretUp: caretUp ?? _this.caretUp,
       checkmark: checkmark ?? _this.checkmark,
       checks: checks ?? _this.checks,
       chevronDown: chevronDown ?? _this.chevronDown,
@@ -272,11 +364,14 @@ mixin _$StreamIcons {
       clock: clock ?? _this.clock,
       command: command ?? _this.command,
       copy: copy ?? _this.copy,
+      copyFill: copyFill ?? _this.copyFill,
+      darkMode: darkMode ?? _this.darkMode,
       delete: delete ?? _this.delete,
       download: download ?? _this.download,
       edit: edit ?? _this.edit,
       emoji: emoji ?? _this.emoji,
       emojiAdd: emojiAdd ?? _this.emojiAdd,
+      emojiAddFill: emojiAddFill ?? _this.emojiAddFill,
       exclamationCircle: exclamationCircle ?? _this.exclamationCircle,
       exclamationCircleFill:
           exclamationCircleFill ?? _this.exclamationCircleFill,
@@ -289,35 +384,57 @@ mixin _$StreamIcons {
       flag: flag ?? _this.flag,
       folder: folder ?? _this.folder,
       folderLarge: folderLarge ?? _this.folderLarge,
+      fullBlurFill: fullBlurFill ?? _this.fullBlurFill,
+      fullscreenFill: fullscreenFill ?? _this.fullscreenFill,
       gallery: gallery ?? _this.gallery,
+      gridFill: gridFill ?? _this.gridFill,
+      gridPixelFill: gridPixelFill ?? _this.gridPixelFill,
       image: image ?? _this.image,
       imageLarge: imageLarge ?? _this.imageLarge,
       info: info ?? _this.info,
+      language: language ?? _this.language,
       leave: leave ?? _this.leave,
+      leftToRight: leftToRight ?? _this.leftToRight,
+      lightMode: lightMode ?? _this.lightMode,
       link: link ?? _this.link,
       location: location ?? _this.location,
       lock: lock ?? _this.lock,
       megaphone: megaphone ?? _this.megaphone,
       mention: mention ?? _this.mention,
+      menu: menu ?? _this.menu,
       messageBubble: messageBubble ?? _this.messageBubble,
       messageBubbleFill: messageBubbleFill ?? _this.messageBubbleFill,
       messageBubbleLarge: messageBubbleLarge ?? _this.messageBubbleLarge,
       messageBubbles: messageBubbles ?? _this.messageBubbles,
+      messageBubblesFill: messageBubblesFill ?? _this.messageBubblesFill,
       messageBubblesLarge: messageBubblesLarge ?? _this.messageBubblesLarge,
       minus: minus ?? _this.minus,
       minusCircle: minusCircle ?? _this.minusCircle,
       more: more ?? _this.more,
+      moreHorizontal: moreHorizontal ?? _this.moreHorizontal,
+      moreVerticalFill: moreVerticalFill ?? _this.moreVerticalFill,
       mute: mute ?? _this.mute,
       noSign: noSign ?? _this.noSign,
+      noSignFill: noSignFill ?? _this.noSignFill,
       notification: notification ?? _this.notification,
       pauseFill: pauseFill ?? _this.pauseFill,
+      phoneDownFill: phoneDownFill ?? _this.phoneDownFill,
       pin: pin ?? _this.pin,
+      pinFill: pinFill ?? _this.pinFill,
       pinLarge: pinLarge ?? _this.pinLarge,
+      pipFill: pipFill ?? _this.pipFill,
       playFill: playFill ?? _this.playFill,
       plus: plus ?? _this.plus,
       poll: poll ?? _this.poll,
       pollLarge: pollLarge ?? _this.pollLarge,
+      presentDesktopFill: presentDesktopFill ?? _this.presentDesktopFill,
+      presentMobileFill: presentMobileFill ?? _this.presentMobileFill,
+      questionCircleFill: questionCircleFill ?? _this.questionCircleFill,
       quote: quote ?? _this.quote,
+      raiseHandFill: raiseHandFill ?? _this.raiseHandFill,
+      recordLibraryFill: recordLibraryFill ?? _this.recordLibraryFill,
+      recordingFill: recordingFill ?? _this.recordingFill,
+      recordingStopFill: recordingStopFill ?? _this.recordingStopFill,
       refresh: refresh ?? _this.refresh,
       reorder: reorder ?? _this.reorder,
       reply: reply ?? _this.reply,
@@ -325,9 +442,14 @@ mixin _$StreamIcons {
       save: save ?? _this.save,
       search: search ?? _this.search,
       send: send ?? _this.send,
+      settings: settings ?? _this.settings,
+      settingsFill: settingsFill ?? _this.settingsFill,
       share: share ?? _this.share,
       shield: shield ?? _this.shield,
       sidebar: sidebar ?? _this.sidebar,
+      slidersFill: slidersFill ?? _this.slidersFill,
+      starFill: starFill ?? _this.starFill,
+      statsFill: statsFill ?? _this.statsFill,
       stopFill: stopFill ?? _this.stopFill,
       thread: thread ?? _this.thread,
       threadFill: threadFill ?? _this.threadFill,
@@ -335,19 +457,26 @@ mixin _$StreamIcons {
       trophy: trophy ?? _this.trophy,
       unlock: unlock ?? _this.unlock,
       unpin: unpin ?? _this.unpin,
+      unpinFill: unpinFill ?? _this.unpinFill,
       unsave: unsave ?? _this.unsave,
       unsupportedAttachment:
           unsupportedAttachment ?? _this.unsupportedAttachment,
       upload: upload ?? _this.upload,
       user: user ?? _this.user,
       userAdd: userAdd ?? _this.userAdd,
+      userAddFill: userAddFill ?? _this.userAddFill,
       userCheck: userCheck ?? _this.userCheck,
       userRemove: userRemove ?? _this.userRemove,
+      userRemoveFill: userRemoveFill ?? _this.userRemoveFill,
       users: users ?? _this.users,
+      usersFill: usersFill ?? _this.usersFill,
+      verifiedFill: verifiedFill ?? _this.verifiedFill,
       video: video ?? _this.video,
       videoFill: videoFill ?? _this.videoFill,
+      videoOffFill: videoOffFill ?? _this.videoOffFill,
       voice: voice ?? _this.voice,
       voiceFill: voiceFill ?? _this.voiceFill,
+      voiceOffFill: voiceOffFill ?? _this.voiceOffFill,
       xCircle: xCircle ?? _this.xCircle,
       xmark: xmark ?? _this.xmark,
       xmarkSmall: xmarkSmall ?? _this.xmarkSmall,
@@ -380,9 +509,15 @@ mixin _$StreamIcons {
       audio: other.audio,
       bell: other.bell,
       bellOff: other.bellOff,
+      blurFill: other.blurFill,
       bolt: other.bolt,
+      boltFill: other.boltFill,
       camera: other.camera,
+      cameraFlipFill: other.cameraFlipFill,
       cameraLarge: other.cameraLarge,
+      captionFill: other.captionFill,
+      caretDown: other.caretDown,
+      caretUp: other.caretUp,
       checkmark: other.checkmark,
       checks: other.checks,
       chevronDown: other.chevronDown,
@@ -392,11 +527,14 @@ mixin _$StreamIcons {
       clock: other.clock,
       command: other.command,
       copy: other.copy,
+      copyFill: other.copyFill,
+      darkMode: other.darkMode,
       delete: other.delete,
       download: other.download,
       edit: other.edit,
       emoji: other.emoji,
       emojiAdd: other.emojiAdd,
+      emojiAddFill: other.emojiAddFill,
       exclamationCircle: other.exclamationCircle,
       exclamationCircleFill: other.exclamationCircleFill,
       exclamationMarkFill: other.exclamationMarkFill,
@@ -407,35 +545,57 @@ mixin _$StreamIcons {
       flag: other.flag,
       folder: other.folder,
       folderLarge: other.folderLarge,
+      fullBlurFill: other.fullBlurFill,
+      fullscreenFill: other.fullscreenFill,
       gallery: other.gallery,
+      gridFill: other.gridFill,
+      gridPixelFill: other.gridPixelFill,
       image: other.image,
       imageLarge: other.imageLarge,
       info: other.info,
+      language: other.language,
       leave: other.leave,
+      leftToRight: other.leftToRight,
+      lightMode: other.lightMode,
       link: other.link,
       location: other.location,
       lock: other.lock,
       megaphone: other.megaphone,
       mention: other.mention,
+      menu: other.menu,
       messageBubble: other.messageBubble,
       messageBubbleFill: other.messageBubbleFill,
       messageBubbleLarge: other.messageBubbleLarge,
       messageBubbles: other.messageBubbles,
+      messageBubblesFill: other.messageBubblesFill,
       messageBubblesLarge: other.messageBubblesLarge,
       minus: other.minus,
       minusCircle: other.minusCircle,
       more: other.more,
+      moreHorizontal: other.moreHorizontal,
+      moreVerticalFill: other.moreVerticalFill,
       mute: other.mute,
       noSign: other.noSign,
+      noSignFill: other.noSignFill,
       notification: other.notification,
       pauseFill: other.pauseFill,
+      phoneDownFill: other.phoneDownFill,
       pin: other.pin,
+      pinFill: other.pinFill,
       pinLarge: other.pinLarge,
+      pipFill: other.pipFill,
       playFill: other.playFill,
       plus: other.plus,
       poll: other.poll,
       pollLarge: other.pollLarge,
+      presentDesktopFill: other.presentDesktopFill,
+      presentMobileFill: other.presentMobileFill,
+      questionCircleFill: other.questionCircleFill,
       quote: other.quote,
+      raiseHandFill: other.raiseHandFill,
+      recordLibraryFill: other.recordLibraryFill,
+      recordingFill: other.recordingFill,
+      recordingStopFill: other.recordingStopFill,
       refresh: other.refresh,
       reorder: other.reorder,
       reply: other.reply,
@@ -443,9 +603,14 @@ mixin _$StreamIcons {
       save: other.save,
       search: other.search,
       send: other.send,
+      settings: other.settings,
+      settingsFill: other.settingsFill,
       share: other.share,
       shield: other.shield,
       sidebar: other.sidebar,
+      slidersFill: other.slidersFill,
+      starFill: other.starFill,
+      statsFill: other.statsFill,
       stopFill: other.stopFill,
       thread: other.thread,
       threadFill: other.threadFill,
@@ -453,18 +618,25 @@ mixin _$StreamIcons {
       trophy: other.trophy,
       unlock: other.unlock,
       unpin: other.unpin,
+      unpinFill: other.unpinFill,
       unsave: other.unsave,
       unsupportedAttachment: other.unsupportedAttachment,
       upload: other.upload,
       user: other.user,
       userAdd: other.userAdd,
+      userAddFill: other.userAddFill,
       userCheck: other.userCheck,
       userRemove: other.userRemove,
+      userRemoveFill: other.userRemoveFill,
       users: other.users,
+      usersFill: other.usersFill,
+      verifiedFill: other.verifiedFill,
       video: other.video,
       videoFill: other.videoFill,
+      videoOffFill: other.videoOffFill,
       voice: other.voice,
       voiceFill: other.voiceFill,
+      voiceOffFill: other.voiceOffFill,
       xCircle: other.xCircle,
       xmark: other.xmark,
       xmarkSmall: other.xmarkSmall,
@@ -498,9 +670,15 @@ mixin _$StreamIcons {
         _other.audio == _this.audio &&
         _other.bell == _this.bell &&
         _other.bellOff == _this.bellOff &&
+        _other.blurFill == _this.blurFill &&
         _other.bolt == _this.bolt &&
+        _other.boltFill == _this.boltFill &&
         _other.camera == _this.camera &&
+        _other.cameraFlipFill == _this.cameraFlipFill &&
         _other.cameraLarge == _this.cameraLarge &&
+        _other.captionFill == _this.captionFill &&
+        _other.caretDown == _this.caretDown &&
+        _other.caretUp == _this.caretUp &&
         _other.checkmark == _this.checkmark &&
         _other.checks == _this.checks &&
         _other.chevronDown == _this.chevronDown &&
@@ -510,11 +688,14 @@ mixin _$StreamIcons {
         _other.clock == _this.clock &&
         _other.command == _this.command &&
         _other.copy == _this.copy &&
+        _other.copyFill == _this.copyFill &&
+        _other.darkMode == _this.darkMode &&
         _other.delete == _this.delete &&
         _other.download == _this.download &&
         _other.edit == _this.edit &&
         _other.emoji == _this.emoji &&
         _other.emojiAdd == _this.emojiAdd &&
+        _other.emojiAddFill == _this.emojiAddFill &&
         _other.exclamationCircle == _this.exclamationCircle &&
         _other.exclamationCircleFill == _this.exclamationCircleFill &&
         _other.exclamationMarkFill == _this.exclamationMarkFill &&
@@ -525,35 +706,57 @@ mixin _$StreamIcons {
         _other.flag == _this.flag &&
         _other.folder == _this.folder &&
         _other.folderLarge == _this.folderLarge &&
+        _other.fullBlurFill == _this.fullBlurFill &&
+        _other.fullscreenFill == _this.fullscreenFill &&
         _other.gallery == _this.gallery &&
+        _other.gridFill == _this.gridFill &&
+        _other.gridPixelFill == _this.gridPixelFill &&
         _other.image == _this.image &&
         _other.imageLarge == _this.imageLarge &&
         _other.info == _this.info &&
+        _other.language == _this.language &&
         _other.leave == _this.leave &&
+        _other.leftToRight == _this.leftToRight &&
+        _other.lightMode == _this.lightMode &&
         _other.link == _this.link &&
         _other.location == _this.location &&
         _other.lock == _this.lock &&
         _other.megaphone == _this.megaphone &&
         _other.mention == _this.mention &&
+        _other.menu == _this.menu &&
         _other.messageBubble == _this.messageBubble &&
         _other.messageBubbleFill == _this.messageBubbleFill &&
         _other.messageBubbleLarge == _this.messageBubbleLarge &&
         _other.messageBubbles == _this.messageBubbles &&
+        _other.messageBubblesFill == _this.messageBubblesFill &&
         _other.messageBubblesLarge == _this.messageBubblesLarge &&
         _other.minus == _this.minus &&
         _other.minusCircle == _this.minusCircle &&
         _other.more == _this.more &&
+        _other.moreHorizontal == _this.moreHorizontal &&
+        _other.moreVerticalFill == _this.moreVerticalFill &&
         _other.mute == _this.mute &&
         _other.noSign == _this.noSign &&
+        _other.noSignFill == _this.noSignFill &&
         _other.notification == _this.notification &&
         _other.pauseFill == _this.pauseFill &&
+        _other.phoneDownFill == _this.phoneDownFill &&
         _other.pin == _this.pin &&
+        _other.pinFill == _this.pinFill &&
         _other.pinLarge == _this.pinLarge &&
+        _other.pipFill == _this.pipFill &&
         _other.playFill == _this.playFill &&
         _other.plus == _this.plus &&
         _other.poll == _this.poll &&
         _other.pollLarge == _this.pollLarge &&
+        _other.presentDesktopFill == _this.presentDesktopFill &&
+        _other.presentMobileFill == _this.presentMobileFill &&
+        _other.questionCircleFill == _this.questionCircleFill &&
         _other.quote == _this.quote &&
+        _other.raiseHandFill == _this.raiseHandFill &&
+        _other.recordLibraryFill == _this.recordLibraryFill &&
+        _other.recordingFill == _this.recordingFill &&
+        _other.recordingStopFill == _this.recordingStopFill &&
         _other.refresh == _this.refresh &&
         _other.reorder == _this.reorder &&
         _other.reply == _this.reply &&
@@ -561,9 +764,14 @@ mixin _$StreamIcons {
         _other.save == _this.save &&
         _other.search == _this.search &&
         _other.send == _this.send &&
+        _other.settings == _this.settings &&
+        _other.settingsFill == _this.settingsFill &&
         _other.share == _this.share &&
         _other.shield == _this.shield &&
         _other.sidebar == _this.sidebar &&
+        _other.slidersFill == _this.slidersFill &&
+        _other.starFill == _this.starFill &&
+        _other.statsFill == _this.statsFill &&
         _other.stopFill == _this.stopFill &&
         _other.thread == _this.thread &&
         _other.threadFill == _this.threadFill &&
@@ -571,18 +779,25 @@ mixin _$StreamIcons {
         _other.trophy == _this.trophy &&
         _other.unlock == _this.unlock &&
         _other.unpin == _this.unpin &&
+        _other.unpinFill == _this.unpinFill &&
         _other.unsave == _this.unsave &&
         _other.unsupportedAttachment == _this.unsupportedAttachment &&
         _other.upload == _this.upload &&
         _other.user == _this.user &&
         _other.userAdd == _this.userAdd &&
+        _other.userAddFill == _this.userAddFill &&
         _other.userCheck == _this.userCheck &&
         _other.userRemove == _this.userRemove &&
+        _other.userRemoveFill == _this.userRemoveFill &&
         _other.users == _this.users &&
+        _other.usersFill == _this.usersFill &&
+        _other.verifiedFill == _this.verifiedFill &&
         _other.video == _this.video &&
         _other.videoFill == _this.videoFill &&
+        _other.videoOffFill == _this.videoOffFill &&
         _other.voice == _this.voice &&
         _other.voiceFill == _this.voiceFill &&
+        _other.voiceOffFill == _this.voiceOffFill &&
         _other.xCircle == _this.xCircle &&
         _other.xmark == _this.xmark &&
         _other.xmarkSmall == _this.xmarkSmall &&
@@ -608,9 +823,15 @@ mixin _$StreamIcons {
       _this.audio,
       _this.bell,
       _this.bellOff,
+      _this.blurFill,
       _this.bolt,
+      _this.boltFill,
       _this.camera,
+      _this.cameraFlipFill,
       _this.cameraLarge,
+      _this.captionFill,
+      _this.caretDown,
+      _this.caretUp,
       _this.checkmark,
       _this.checks,
       _this.chevronDown,
@@ -620,11 +841,14 @@ mixin _$StreamIcons {
       _this.clock,
       _this.command,
       _this.copy,
+      _this.copyFill,
+      _this.darkMode,
       _this.delete,
       _this.download,
       _this.edit,
       _this.emoji,
       _this.emojiAdd,
+      _this.emojiAddFill,
       _this.exclamationCircle,
       _this.exclamationCircleFill,
       _this.exclamationMarkFill,
@@ -635,35 +859,57 @@ mixin _$StreamIcons {
       _this.flag,
       _this.folder,
       _this.folderLarge,
+      _this.fullBlurFill,
+      _this.fullscreenFill,
       _this.gallery,
+      _this.gridFill,
+      _this.gridPixelFill,
       _this.image,
       _this.imageLarge,
       _this.info,
+      _this.language,
       _this.leave,
+      _this.leftToRight,
+      _this.lightMode,
       _this.link,
       _this.location,
       _this.lock,
       _this.megaphone,
       _this.mention,
+      _this.menu,
       _this.messageBubble,
       _this.messageBubbleFill,
       _this.messageBubbleLarge,
       _this.messageBubbles,
+      _this.messageBubblesFill,
       _this.messageBubblesLarge,
       _this.minus,
       _this.minusCircle,
       _this.more,
+      _this.moreHorizontal,
+      _this.moreVerticalFill,
       _this.mute,
       _this.noSign,
+      _this.noSignFill,
       _this.notification,
       _this.pauseFill,
+      _this.phoneDownFill,
       _this.pin,
+      _this.pinFill,
       _this.pinLarge,
+      _this.pipFill,
       _this.playFill,
       _this.plus,
       _this.poll,
       _this.pollLarge,
+      _this.presentDesktopFill,
+      _this.presentMobileFill,
+      _this.questionCircleFill,
       _this.quote,
+      _this.raiseHandFill,
+      _this.recordLibraryFill,
+      _this.recordingFill,
+      _this.recordingStopFill,
       _this.refresh,
       _this.reorder,
       _this.reply,
@@ -671,9 +917,14 @@ mixin _$StreamIcons {
       _this.save,
       _this.search,
       _this.send,
+      _this.settings,
+      _this.settingsFill,
       _this.share,
       _this.shield,
       _this.sidebar,
+      _this.slidersFill,
+      _this.starFill,
+      _this.statsFill,
       _this.stopFill,
       _this.thread,
       _this.threadFill,
@@ -681,18 +932,25 @@ mixin _$StreamIcons {
       _this.trophy,
       _this.unlock,
       _this.unpin,
+      _this.unpinFill,
       _this.unsave,
       _this.unsupportedAttachment,
       _this.upload,
       _this.user,
       _this.userAdd,
+      _this.userAddFill,
       _this.userCheck,
       _this.userRemove,
+      _this.userRemoveFill,
       _this.users,
+      _this.usersFill,
+      _this.verifiedFill,
       _this.video,
       _this.videoFill,
+      _this.videoOffFill,
       _this.voice,
       _this.voiceFill,
+      _this.voiceOffFill,
       _this.xCircle,
       _this.xmark,
       _this.xmarkSmall,